# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quetzal API
  version: "1.0"
  license:
    name: MIT
  description: |
    HTTP API for Quetzal, a Kubernetes-native game-server control plane.

    ## Authentication
    Most endpoints require authentication. Two credential types are accepted:

    - **API key** (recommended for scripts): send `Authorization: Bearer <token>`.
      Create a key under *Account → API keys*; it inherits the owner's
      permissions. Tokens are shown once at creation.
    - **Session cookie**: set by `POST /api/login` and used by the web UI.

    A key/session identifies a user; per-server access follows that user's
    ownership and granted subuser permissions. Admin-only endpoints are noted.

    ## Conventions
    - Request and response bodies are JSON unless noted.
    - Mutating requests made from a browser are subject to a same-origin check.
    - Authentication endpoints are rate limited (HTTP 429 with `Retry-After`).
servers:
  - url: /
    description: Same origin as the panel
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/servers/{id}/backups:
    parameters:
      - $ref: "#/components/parameters/ServerId"
    get:
      tags: [Backups]
      summary: List backups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Backup" }
    post:
      tags: [Backups]
      summary: Trigger a backup
      responses:
        "201":
          description: Backup started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Backup" }
...
  /api/servers/{id}/backups/{bid}:
    parameters:
      - $ref: "#/components/parameters/ServerId"
      - { name: bid, in: path, required: true, schema: { type: integer } }
    delete:
      tags: [Backups]
      summary: Delete a backup
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key created under Account → API keys.
    cookieAuth:
      type: apiKey
      in: cookie
      name: quetzal_session
      description: Session cookie set by POST /api/login (used by the web UI).
  parameters:
    ServerId:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: Server ID
...
  schemas:
...
    Backup:
      type: object
      properties:
        id: { type: integer }
        serverId: { type: integer }
        direction: { type: string, enum: [backup, restore] }
        phase: { type: string, enum: [Pending, Running, Succeeded, Failed] }
        sourceId: { type: integer }
        sizeBytes: { type: integer, format: int64 }
        message: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
...